MATH-1231
Lifted unnecessary type restriction on constructor's "exponent" argument
of "MicrosphereInterpolator" and "MicrosphereInterpolatingFunction".

+++ src/main/java/org/apache/commons/math4/analysis/interpolation/MicrosphereInterpolatingFunction.java
@@ -144,7 +144,7 @@ public class MicrosphereInterpolatingFunction
      */
     public MicrosphereInterpolatingFunction(double[][] xval,
                                             double[] yval,
-                                            int brightnessExponent,
+                                            double brightnessExponent,
                                             int microsphereElements,
                                             UnitSphereRandomVectorGenerator rand)
         throws DimensionMismatchException,
